test: VertexColorRoundtrip のレビュー指摘を修正
- 壊れたprogramの判定を diagnostics の有無から diagnostics.runnable === false に変更。
  three の WebGLProgram はリンク成功時でも info log が空でなければ
  diagnostics を生成する（runnable: true）ため、コンパイラの警告だけで
  テストが落ちうる状態だった。特にこの修正は vColor を
  「書かれるが読まれない varying」にするため警告が出やすい
- renderAndInspect を beforeAll で1回だけ実行するようにした。
  SwiftShader でのフルレンダリングが2回走り、WebGLコンテキストも
  2つ確保していた。あわせて forceContextLoss() でコンテキストを解放する
  （dispose() だけでは解放されない）
- loadVRM / exportVRM が MToonAtlasRoundtrip.test.ts と逐語重複していたため
  tests/browser/helpers/vrm-roundtrip.ts に共通化し、既存テストも差し替えた
